refactor item scoring
